Maryland/requirements.txt
# Core dependencies
supabase==2.0.0
pandas==2.1.0
pydantic==2.0.0
python-dotenv==1.0.0

# HTTP and networking
requests==2.31.0
tenacity==8.2.3

# Data processing
python-dateutil==2.8.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1

# Optional for development/testing
pytest==7.4.0
pytest-asyncio==0.21.0
black==23.7.0
flake8==6.0.0